chore: move MongoDb code to legacy package (Publications)
