+++ src/main/kotlin/fr/nicopico/n2rss/utils/CleanLocalDatabase.kt
@@ -18,7 +18,7 @@
 
 package fr.nicopico.n2rss.utils
 
-import fr.nicopico.n2rss.newsletter.data.PublicationRepository
+import fr.nicopico.n2rss.newsletter.data.legacy.LegacyPublicationRepository
 import org.slf4j.LoggerFactory
 import org.springframework.context.annotation.Profile
 import org.springframework.context.event.ContextRefreshedEvent
@@ -28,12 +28,12 @@ import org.springframework.stereotype.Component
 @Profile("local")
 @Component
 class CleanLocalDatabase(
-    private val publicationRepository: PublicationRepository,
+    private val legacyPublicationRepository: LegacyPublicationRepository,
 ) {
     @EventListener
     fun onApplicationEvent(ignored: ContextRefreshedEvent) {
         LOG.info("Clean-up local database...")
-        publicationRepository.deleteAll()
+        legacyPublicationRepository.deleteAll()
     }
 
     companion object {
