+++ src/test/kotlin/fr/nicopico/n2rss/utils/CleanLocalDatabaseTest.kt
@@ -18,7 +18,7 @@
 
 package fr.nicopico.n2rss.utils
 
-import fr.nicopico.n2rss.newsletter.data.PublicationRepository
+import fr.nicopico.n2rss.newsletter.data.legacy.LegacyPublicationRepository
 import io.mockk.MockKAnnotations
 import io.mockk.confirmVerified
 import io.mockk.impl.annotations.MockK
@@ -31,7 +31,7 @@ import org.springframework.context.event.ContextRefreshedEvent
 class CleanLocalDatabaseTest {
 
     @MockK(relaxUnitFun = true)
-    private lateinit var publicationRepository: PublicationRepository
+    private lateinit var legacyPublicationRepository: LegacyPublicationRepository
 
     private lateinit var cleanLocalDatabase: CleanLocalDatabase
 
@@ -39,7 +39,7 @@ class CleanLocalDatabaseTest {
     fun setUp() {
         MockKAnnotations.init(this)
         cleanLocalDatabase = CleanLocalDatabase(
-            publicationRepository,
+            legacyPublicationRepository,
         )
     }
 
@@ -53,10 +53,10 @@ class CleanLocalDatabaseTest {
 
         // THEN
         verify(exactly = 1) {
-            publicationRepository.deleteAll()
+            legacyPublicationRepository.deleteAll()
         }
         confirmVerified(
-            publicationRepository,
+            legacyPublicationRepository,
         )
     }
 }
